Force setuptools to use distro's distutils
On the Ubuntu/Debian-packaged Python distribution, distutils is patched to
understand an option `setup.py install --install-layout=deb`. This option is
used by catkin setup.py scripts.

When we installed a newer setuptools (830f301), it came with a newer, vendored
version of distutils which does not have the Ubuntu/Debian patch, and hence
doesn't work with catkin.

By setting the SETUPTOOLS_USE_DISTUTILS=stdlib environment variable, we force
setuptools to use the distro-packaged distutils with the necessary feature.

This was never a problem before when we ran `catkin build` because it did not
actually invoke the `install` command, it did some dry run shenanigans.

Python version constraint: distutils was removed from Python's stdlib in
version 3.12 (PEP 632). This means we cannot upgrade past Python 3.11 while
continuing to support ROS Noetic. (Ubuntu 24.04 ships Python 3.12.)